% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_info_reference}
\alias{api_info_reference}
\title{Geocoding service links and information}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoding service name}
\item{method_display_name}{Geocoding service display name}
\item{site_url}{Link to the main site of the geocoding service}
\item{api_documentation_url}{Link to API documentation}
\item{api_usage_policy_url}{Link to the usage policy}
}
}
\usage{
api_info_reference
}
\description{
This dataset is used for generating package documentation.
}
\keyword{datasets}
