% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_key_reference}
\alias{api_key_reference}
\title{The name of the environmental variable that the API key will be read from.}
\format{
A tibble dataframe
\describe{
 \item{method}{Geocoder service name}
 \item{env_var}{Environmental variable name}
}
}
\usage{
api_key_reference
}
\description{
The name of the environmental variable that the API key will be read from.
}
\keyword{datasets}
