% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobatch.R
\name{batch_census}
\alias{batch_census}
\title{Census batch geocoding
Vingate must be defined if return = 'geographies'}
\usage{
batch_census(address_list, return = "locations", timeout = 5)
}
\description{
Census batch geocoding
Vingate must be defined if return = 'geographies'
}
