% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_geocoding.R
\name{batch_geocodio}
\alias{batch_geocodio}
\title{Batch geocode with geocodio}
\usage{
batch_geocodio(
  address_pack,
  lat = "lat",
  long = "long",
  timeout = 5,
  full_results = FALSE,
  ...
)
}
\description{
Batch geocode with geocodio
}
