% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract latitude, longitude as as numeric vector c(lat, long)}
\usage{
extract_coords(method, response)
}
\arguments{
\item{method}{method name}

\item{response}{content from the geocoder service}
}
\value{
latitude and longitude as a numeric vector
}
\description{
Extract latitude, longitude as as numeric vector c(lat, long)
}
