% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_results}
\alias{extract_results}
\title{Extract geocoder results (exclude lat/long coordinates)}
\usage{
extract_results(method, response, flatten = TRUE)
}
\arguments{
\item{method}{method name}

\item{response}{content from the geocoder service}

\item{flatten}{if TRUE then}
}
\value{
results in dataframe format
}
\description{
Extract geocoder results (exclude lat/long coordinates)
}
