% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo}
\alias{geo}
\title{Geocode addresses}
\usage{
geo(
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  method = "osm",
  lat = "lat",
  long = "long",
  limit = 1,
  full_results = FALSE,
  mode = "",
  unique_only = FALSE,
  return_addresses = TRUE,
  min_time = NULL,
  progress_bar = show_progress_bar(),
  quiet = getOption("tidygeocoder.quiet", FALSE),
  api_url = NULL,
  timeout = 20,
  flatten = TRUE,
  batch_limit = NULL,
  verbose = getOption("tidygeocoder.verbose", FALSE),
  no_query = FALSE,
  custom_query = list(),
  api_options = list()
)
}
\arguments{
\item{address}{single line address (ie. '1600 Pennsylvania Ave NW, Washington, DC').
Do not combine with the address component arguments below
(\code{street}, \code{city}, \code{county}, \code{state}, \code{postalcode}, \code{country}).}

\item{street}{street address (ie. '1600 Pennsylvania Ave NW')}

\item{city}{city (ie. 'Tokyo')}

\item{county}{county (ie. 'Jefferson')}

\item{state}{state (ie. 'Kentucky')}

\item{postalcode}{postalcode (ie. zip code if in the United States)}

\item{country}{country (ie. 'Japan')}

\item{method}{the geocoding service to be used. API keys are loaded from environmental variables. Run \code{usethis::edit_r_environ()} to open your .Renviron file and add an API key as an environmental variable. For example, add the line \code{GEOCODIO_API_KEY="YourAPIKeyHere"}
\itemize{
\item \code{"osm"}: \href{https://nominatim.org}{Nominatim}.
\item \code{"census"}: \href{https://geocoding.geo.census.gov/}{US Census}. Geographic coverage is limited to the United States. Batch geocoding is supported.
\item \code{"arcgis"}: \href{https://developers.arcgis.com/rest/geocode/api-reference/overview-world-geocoding-service.htm}{ArcGIS}.
\item \code{"geocodio"}: \href{https://www.geocod.io/}{Geocodio}. Geographic coverage is limited to the United States and Canada. Batch geocoding is supported. Store an API key in the environmental variable \code{"GEOCODIO_API_KEY"}.
\item \code{"iq"}: \href{https://locationiq.com/}{Location IQ}.   Store an API key in the environmental variable \code{"LOCATIONIQ_API_KEY"}.
\item \code{"google"}: \href{https://developers.google.com/maps/documentation/geocoding/overview}{Google}.   Store an API key in the environmental variable \code{"GOOGLEGEOCODE_API_KEY"}.
\item \code{"opencage"}: \href{https://opencagedata.com}{OpenCage}.   Store an API key in the environmental variable \code{"OPENCAGE_KEY"}.
\item \code{"mapbox"}: \href{https://docs.mapbox.com/api/search/}{Mapbox}.   Store an API key in the environmental variable \code{"MAPBOX_API_KEY"}.
\item \code{"here"}: \href{https://www.here.com/platform/geocoding}{HERE}.  Batch geocoding is supported, but must be explicitly called with \code{mode = "batch"}. Store an API key in the environmental variable \code{"HERE_API_KEY"}.
\item \code{"tomtom"}: \href{https://developer.tomtom.com/search-api/search-api-documentation}{TomTom}.  Batch geocoding is supported. Store an API key in the environmental variable \code{"TOMTOM_API_KEY"}.
\item \code{"mapquest"}: \href{https://developer.mapquest.com/documentation/api/geocoding}{MapQuest}.  Batch geocoding is supported. Store an API key in the environmental variable \code{"MAPQUEST_API_KEY"}.
\item \code{"bing"}: \href{https://learn.microsoft.com/en-us/bingmaps/rest-services/locations/}{Bing}.  Batch geocoding is supported, but must be explicitly called with \code{mode = "batch"}. Store an API key in the environmental variable \code{"BINGMAPS_API_KEY"}.
\item \code{"geoapify"}: \href{https://www.geoapify.com/geocoding-api}{Geoapify}.   Store an API key in the environmental variable \code{"GEOAPIFY_KEY"}.
\item \code{"vietmap"}: \href{https://maps.vietmap.vn/docs/}{VietMap}.   Store an API key in the environmental variable \code{"VIETMAP_API_KEY"}.
}}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. \code{lat} or \code{"lat"}).}

\item{long}{longitude column name. Can be quoted or unquoted (ie. \code{long} or \code{"long"}).}

\item{limit}{maximum number of results to return per input address. For many geocoding services the maximum value of the limit parameter is 100. Pass \code{limit = NULL} to use the default \code{limit} value of the selected geocoding service. For batch geocoding, limit must be set to 1 (default) if \code{return_addresses = TRUE}. Refer to \link{api_parameter_reference} for more details.}

\item{full_results}{returns all available data from the geocoding service if TRUE. If FALSE (default) then only latitude and longitude columns are returned from the geocoding service.}

\item{mode}{set to 'batch' to force batch geocoding or 'single' to force single address geocoding (one address per query). If not specified then batch geocoding will be used if available (given method selected) when multiple addresses are provided; otherwise single address geocoding will be used. For the "here" and "bing" methods the batch mode should be explicitly specified with \code{mode = 'batch'}.}

\item{unique_only}{only return results for unique inputs if TRUE}

\item{return_addresses}{return input addresses with results if TRUE. Note that
most services return the input addresses with \code{full_results = TRUE} and setting
return_addresses to FALSE does not prevent this.}

\item{min_time}{minimum amount of time for a query to take (in seconds). If NULL
then min_time will be set to the default value specified in \link{min_time_reference}.}

\item{progress_bar}{if TRUE then a progress bar will be displayed
for single input geocoding (1 input per query). By default the progress bar
will not be shown for code executed when knitting R Markdown files or code within
an RStudio notebook chunk. Can be set permanently with \code{options(tidygeocoder.progress_bar = FALSE)}.}

\item{quiet}{if TRUE then console messages that are displayed by default
regarding queries will be suppressed. FALSE is default.
Can be set permanently with \code{options(tidygeocoder.quiet = TRUE)}.}

\item{api_url}{custom API URL. If specified, the default API URL will be overridden.
This parameter can be used to specify a local Nominatim server, for instance.}

\item{timeout}{query timeout (in minutes)}

\item{flatten}{if TRUE (default) then any nested dataframes in results are flattened if possible.
Note that in some cases results are flattened regardless such as for Geocodio batch geocoding.}

\item{batch_limit}{limit to the number of addresses in a batch geocoding query.Defaults to the value in \link{batch_limit_reference} if not specified.}

\item{verbose}{if TRUE then detailed logs are output to the console. FALSE is default. Can be set
permanently with \code{options(tidygeocoder.verbose = TRUE)}}

\item{no_query}{if TRUE then no queries are sent to the geocoding service and verbose is set to TRUE.
Used for testing.}

\item{custom_query}{API-specific parameters to be used, passed as a named list
(ex. \code{list(extratags = 1)}.}

\item{api_options}{a named list of parameters specific to individual services.
(ex. \code{list(geocodio_v = 1.6, geocodio_hipaa = TRUE)}). Each parameter begins
with the name of the \code{method} (service) it applies to. The possible parameters
are shown below with their default values.
\itemize{
\item \code{census_return_type} (default: locations): set to \code{"geographies"} to return
additional geography columns. Make sure to use \code{full_results = TRUE} if using
the "geographies" setting.
\item \code{iq_region} (default: "us"): set to "eu" to use the European Union API endpoint
\item \code{geocodio_v} (default: 1.7): the version number of the Geocodio API to be used
\item \code{geocodio_hipaa} (default: FALSE): set to \code{TRUE} to use the HIPAA compliant
Geocodio API endpoint
\item \code{mapbox_permanent} (default: FALSE): set to \code{TRUE} to use the \code{mapbox.places-permanent}
endpoint. Note that this option should be used only if you have applied for a permanent
account. Unsuccessful requests made by an account that does not have access to the
endpoint may be billable.
\item \code{mapquest_open} (default: FALSE): set to \code{TRUE} to use the Open Geocoding endpoint which
relies solely on OpenStreetMap data
\item \code{here_request_id}: this parameter would return a previous HERE batch job,
identified by its RequestID. The RequestID of a batch job is displayed
when \code{verbose} is TRUE. Note that this option would ignore the
current \code{address} parameter on the request, so the \code{return_addresses} or \code{return_coords}
parameters need to be FALSE.
\item \code{vietmap_display_type} (default: FALSE): specify the administrative format to
use, more information at https://maps.vietmap.vn/docs/map-api/geocode-version/geocode-v4/#display-type-options
}}
}
\value{
tibble (dataframe)
}
\description{
Geocodes addresses given as character values. The \link{geocode}
function utilizes this function on addresses contained in dataframes.
See example usage in \code{vignette("tidygeocoder")}.

Note that not all geocoding services support certain address component
parameters. For example, the Census geocoder only covers the United States
and does not have a "country" parameter.

Refer to \link{api_parameter_reference},
\link{min_time_reference}, and \link{batch_limit_reference} for more details on
geocoding service parameters and usage.

This function uses the \link{get_api_query}, \link{query_api}, and
\link{extract_results} functions to create, execute, and parse geocoder
API queries.
}
\examples{
\donttest{
options(tidygeocoder.progress_bar = FALSE)

geo(
  street = "600 Peachtree Street NE", city = "Atlanta",
  state = "Georgia", method = "census"
)

geo(
  address = c("Tokyo, Japan", "Lima, Peru", "Nairobi, Kenya"),
  method = "osm"
)

geo("100 Main St New York, NY",
  full_results = TRUE,
  method = "census", api_options = list(census_return_type = "geographies")
)

geo(
  county = "Jefferson", state = "Kentucky", country = "US",
  method = "osm"
)
}
}
\seealso{
\link{geocode} \link{api_parameter_reference} \link{min_time_reference} \link{batch_limit_reference}
}
