% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_methods.R
\name{geo_cascade}
\alias{geo_cascade}
\title{First try one method and then try another method}
\usage{
geo_cascade(..., cascade_order = c("geocodio", "osm"))
}
\arguments{
\item{...}{arguments passed from and to the \code{\link{geo}} function}

\item{cascade_order}{a vector with two character method values showing 
the order to be attempted}
}
\description{
First try one method and then try another method
}
