% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_combine.R
\name{geo_combine}
\alias{geo_combine}
\title{Combine multiple geocoding queries}
\usage{
geo_combine(
  queries,
  global_params = list(),
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  lat = lat,
  long = long,
  ...
)
}
\arguments{
\item{queries}{list of lists parameter. Each list contains parameters for a query.
The queries are executed in the order provided.
(ie. \code{list(list(method = 'osm'), list(method = 'census'), ...)})}

\item{global_params}{list parameter. Contains parameters to be used for all queries.
(ie. \code{list(address = 'address', full_results = TRUE)})}

\item{address}{single line address (ie. '1600 Pennsylvania Ave NW, Washington, DC').
Do not combine with the address component arguments below
(street, city, county, state, postalcode, country).}

\item{street}{street address (ie. '1600 Pennsylvania Ave NW')}

\item{city}{city (ie. 'Tokyo')}

\item{county}{county (ie. 'Jefferson')}

\item{state}{state (ie. 'Kentucky')}

\item{postalcode}{postalcode (zip code if in the United States)}

\item{country}{country (ie. 'Japan')}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. lat or "lat").}

\item{long}{longitude column name. Can be quoted or unquoted (ie. long or "long").}

\item{...}{arguments passed to the \link{geocode_combine} function}
}
\value{
tibble (dataframe)
}
\description{
Passes address inputs in character vector form to the
\link{geocode_combine} function for geocoding. See example usage in \code{vignette("tidygeocoder")}.

Note that address inputs must be specified for queries either in \code{queries} (for each query)
or \code{global_params} (for all queries) using standard address parameter
names. For example \code{global_params = list(address = 'address')} would
pass addresses from the \code{address} parameter to all queries.
}
\examples{
\donttest{

example_addresses <- c("100 Main St New York, NY", "Paris", "Not a Real Address")

geo_combine(queries = list(list(method = 'census'), list(method = 'osm')),
  address = example_addresses,
  global_params = list(address = 'address')
  )
  
geo_combine(queries = list(
  list(method = 'arcgis'), 
  list(method = 'census', mode = 'single'),
  list(method = 'census', mode = 'batch')
  ),
  global_params = list(address = 'address'),
  address = example_addresses,
  cascade = FALSE,
  return_list = TRUE
)
}
}
\seealso{
\link{geocode_combine} \link{geo} \link{geocode}
}
