% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_osm.R
\name{geo_osm}
\alias{geo_osm}
\title{Geocode addresses with Nomaninatim (OSM)}
\usage{
geo_osm(address, verbose = FALSE, lat = lat, long = long)
}
\arguments{
\item{address}{Single line address}

\item{verbose}{logical. If TRUE outputs logs.}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}
}
\value{
Latitude and Longitude Coordinates in Tibble format
}
\description{
Obtains latitude and longitude coordinates from an address using
the Nominatim (OSM) Geocoder API. Can be used with non-US
or non-street level addresses unlike the Census geocoder
}
\examples{
geo_osm("1600 Pennsylvania Ave Washington, DC")
geo_osm("Paris, France")
}
