% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode a street address}
\usage{
geocode(.tbl, address, method = "Census", verbose = FALSE)
}
\arguments{
\item{.tbl}{Dataframe}

\item{address}{Single line address. Street must be included.}

\item{method}{What geocoder api do you want to use? 'OSM' or 'Census'}

\item{verbose}{logical. If TRUE outputs logs.}
}
\value{
Latitude and Longitude Coordinates in tibble format
}
\description{
Returns latitude and longitude as two columns appended to
the inputted dataframe
}
\examples{

}
