% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{geocode_loop}
\alias{geocode_loop}
\title{New function to replace method = 'cascade'}
\usage{
geocode_loop(
  .tbl,
  queries,
  common_params = list(),
  lat = lat,
  long = long,
  stack = TRUE,
  cascade = FALSE
)
}
\description{
New function to replace method = 'cascade'
}
