% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_handling.R
\name{package_addresses}
\alias{package_addresses}
\title{Function for packaging and deduping addresses}
\usage{
package_addresses(
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  verbose = FALSE
)
}
\description{
Function for packaging and deduping addresses
}
