% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_osm.R
\name{query_osm}
\alias{query_osm}
\title{query_osm}
\usage{
query_osm(address, server = "http://nominatim.openstreetmap.org")
}
\arguments{
\item{address}{- single line address.}

\item{server}{- url to OSM nominatim.}
}
\value{
list containing latitude and longitude coordinates.
}
\description{
Query OSM for an address's latatude and longitude coordinates.
Adapted from tmaptools::geocode_OSM.
}
\references{
\url{https://github.com/mtennekes/tmaptools}
}
