% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{tidygeocoder: Tidyverse-style interface for geocoding.}
\description{
Utilities for geocoding (obtaining latitude and longitude from addresses).
}
\details{
Functions for geocoding (obtaining latitudes and longitude
coordinates from addreses). US Census Geocoder and OSM are supported.

The primary function to use is \code{\link{geocode}} which can call either
\code{\link{geo_census}} or \code{\link{geo_osm}} depending on the method argument
}
\keyword{internal}
