% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_key_reference}
\alias{api_key_reference}
\title{API Key Environmental Variables}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoder service name}
\item{env_var}{Environmental variable name}
}
}
\usage{
api_key_reference
}
\description{
API Key Environmental Variables
}
\details{
API keys are obtained from environmental variables.
The name of the environmental variable used for each service
is stored in this dataset.
}
\seealso{
\link{geo} \link{reverse_geo}
}
\keyword{datasets}
