% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_key_reference}
\alias{api_key_reference}
\title{API key environmental variables}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoding service name}
\item{env_var}{Environmental variable name}
}
}
\usage{
api_key_reference
}
\description{
API keys are obtained from environmental variables.
The \link{geo} and \link{reverse_geo} functions use this dataset
to know which environmental variable to use for
each geocoding service.
}
\seealso{
\link{geo} \link{reverse_geo}
}
\keyword{datasets}
