% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_parameter_reference}
\alias{api_parameter_reference}
\title{REST API Parameter Reference}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoder service shortname}
\item{generic_name}{Generic name of the API parameter}
\item{api_name}{Name of the parameter for the geocoder service specified in the method field}
\item{default_value}{Default value of this parameter}
\item{required}{Is the parameter required by the API?}
}
}
\source{
\itemize{ 
    \item \href{https://locationiq.com/docs}{IQ API Reference} 
    \item \href{https://nominatim.org/release-docs/develop/api/Search}{Nominatim/OSM API Reference}
    \item \href{https://geocoding.geo.census.gov/geocoder/Geocoding_Services_API.pdf}{Census API Documentation}
}
}
\usage{
api_parameter_reference
}
\description{
REST API Parameter Reference
}
\keyword{datasets}
