% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_parameter_reference}
\alias{api_parameter_reference}
\title{Geocoder Service REST API Parameter Reference}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoder service name.}
\item{generic_name}{The universal name of the API parameter}
\item{api_name}{Name of the parameter for the specified geocoder service (method column value)}
\item{default_value}{Default value of this parameter}
\item{required}{Is the parameter required by the specified geocoder service}
}
}
\source{
API Documentation:
\itemize{ 
    \item \href{https://locationiq.com/docs}{LocationIQ (IQ)} 
    \item \href{https://nominatim.org/release-docs/develop/api/Search}{Nominatim (OSM)}
    \item \href{https://geocoding.geo.census.gov/geocoder/Geocoding_Services_API.pdf}{Census}
    \item \href{https://www.geocod.io/docs/}{Geocodio}
}
}
\usage{
api_parameter_reference
}
\description{
This dataset contains the mapping that allows this package to use a 
universal syntax for specifying parameters for different geocoder
service APIs. Reference the documentation for \code{\link{geo}} for
more information.
}
\seealso{
\code{\link{get_api_query}} \code{\link{query_api}}
}
\keyword{datasets}
