% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobatch.R
\name{batch_census}
\alias{batch_census}
\title{Census batch geocoding}
\usage{
batch_census(
  address_pack,
  return = "locations",
  timeout = 15,
  full_results = FALSE,
  verbose = FALSE,
  custom_query = list(),
  lat = "lat",
  long = "long",
  ...
)
}
\arguments{
\item{address_pack}{packaged addresses object}

\item{return}{should be 'locations' or 'geographies'
Vingate must be defined if return = 'geographies'}
}
\description{
Census batch geocoding
}
