% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geobatch.R
\name{batch_census}
\alias{batch_census}
\title{Census batch geocoding
Vingate must be defined if return = 'geographies'}
\usage{
batch_census(
  address = NA,
  street = NA,
  city = NA,
  state = NA,
  postalcode = NA,
  return = "locations",
  timeout = 15,
  full_results = FALSE,
  verbose = FALSE,
  custom_query = list(),
  lat = "lat",
  long = "long"
)
}
\description{
Census batch geocoding
Vingate must be defined if return = 'geographies'
}
