% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_geocoding.R
\name{batch_census}
\alias{batch_census}
\title{Census batch geocoding}
\usage{
batch_census(
  address_pack,
  return = "locations",
  timeout = 20,
  full_results = FALSE,
  custom_query = list(),
  api_url = NULL,
  lat = "lat",
  long = "long",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{address_pack}{packaged addresses object}

\item{return}{should be 'locations' or 'geographies'
Vingate must be defined if return = 'geographies'}
}
\description{
Census batch geocoding
}
