% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_geocoding.R
\name{batch_geocodio}
\alias{batch_geocodio}
\title{Batch geocoding with geocodio}
\usage{
batch_geocodio(
  address_pack,
  lat = "lat",
  long = "long",
  timeout = 20,
  full_results = FALSE,
  verbose = FALSE,
  api_url = NULL,
  geocodio_v = 1.6,
  ...
)
}
\description{
Batch geocoding with geocodio
}
