% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_limit_reference}
\alias{batch_limit_reference}
\title{The batch limit for each geocoder service}
\format{
A tibble dataframe
\describe{
 \item{method}{Geocoder service name}
 \item{batch_limit}{The maximum number of addresses or coordinates allowed per batch}
}
}
\usage{
batch_limit_reference
}
\description{
The batch limit for each geocoder service
}
\keyword{datasets}
