% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_limit_reference}
\alias{batch_limit_reference}
\title{Geocoding batch size limits}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoding service name}
\item{batch_limit}{The maximum number of addresses or coordinates allowed per batch}
}
}
\usage{
batch_limit_reference
}
\description{
The \link{geo} and \link{reverse_geo} functions use this dataset to set the
maximum batch query size for each service.
}
\seealso{
\link{geo} \link{reverse_geo}
}
\keyword{datasets}
