% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract latitude, longitude as as numeric vector c(lat, long)
response is the parsed json response}
\usage{
extract_coords(method, response)
}
\description{
Extract latitude, longitude as as numeric vector c(lat, long)
response is the parsed json response
}
