% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_results}
\alias{extract_results}
\title{Extract geocoder results 
Not used for batch geocoding. Latitude and longitude and extracted into
the first two columns of the returned dataframe}
\usage{
extract_results(method, response, full_results = TRUE, flatten = TRUE)
}
\arguments{
\item{method}{method name}

\item{response}{content from the geocoder service (returned by the )}

\item{flatten}{if TRUE then flatten any nested dataframe content}
}
\description{
Extract geocoder results 
Not used for batch geocoding. Latitude and longitude and extracted into
the first two columns of the returned dataframe
}
