% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_results}
\alias{extract_results}
\title{Extract Geocoder Results}
\usage{
extract_results(method, response, full_results = TRUE, flatten = TRUE)
}
\arguments{
\item{method}{method name}

\item{response}{content from the geocoder service (returned by the \code{\link{query_api}} function)}

\item{full_results}{if TRUE then the full results (not just latitude and longitude)
will be returned.}

\item{flatten}{if TRUE then flatten any nested dataframe content}
}
\value{
geocoder results in tibble format
}
\description{
Parses the output of the \code{\link{query_api}} function.
Latitude and longitude are extracted into the first two columns
of the returned dataframe. This function is not used for batch 
geocoded results.
}
\seealso{
\code{\link{get_api_query}} \code{\link{query_api}} \code{\link{geo}}
}
