% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo}
\alias{geo}
\title{Workhorse function for geocoding}
\usage{
geo(
  method,
  address = NULL,
  api_key = NULL,
  limit = 1,
  api_url = NULL,
  full_results = FALSE,
  verbose = FALSE,
  min_time = NULL
)
}
\arguments{
\item{method}{}

\item{address}{api query parameters in the form of a named list}
}
\value{
parsed results from geocoder
}
\description{
Workhorse function for geocoding
}
