% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo}
\alias{geo}
\title{Workhorse function for geocoding}
\usage{
geo(
  method = "census",
  address = NULL,
  lat = lat,
  long = long,
  api_key = NULL,
  limit = 1,
  api_url = NULL,
  full_results = FALSE,
  verbose = FALSE,
  min_time = NULL,
  debug = FALSE
)
}
\arguments{
\item{method}{the geocoder function you want to use
\itemize{
  \item "census": US Census Geocoder. US street-level addresses only.
  \item "osm": Nominatim (OSM). Worldwide coverage.
}}

\item{address}{address to be geocoded}
}
\value{
parsed results from geocoder
}
\description{
Workhorse function for geocoding
}
