% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{geo}
\alias{geo}
\title{Workhorse function for geocoding}
\usage{
geo(
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  method = "census",
  lat = lat,
  long = long,
  limit = 1,
  api_url = NULL,
  return = "locations",
  vintage = "",
  benchmark = "4",
  custom_query = list(),
  full_results = FALSE,
  verbose = FALSE,
  min_time = NULL
)
}
\arguments{
\item{address}{single line address

Address components (do not combine with the single line 'address' parameter)}

\item{street}{street address}

\item{country}{}

\item{method}{the geocoder function you want to use
\itemize{
  \item "census": US Census Geocoder. US street-level addresses only.
  \item "osm": Nominatim (OSM). Worldwide coverage.
}}

\item{api_url}{Custom URL to use for API. Overrides default URL}

\item{return}{(census only) 'locations' (default) or 'geographies'}

\item{vintage}{(census only)}

\item{benchmark}{(census only)}

\item{custom_query}{API-specific parameters to be used}

\item{full_results}{returns all data from API if True}
}
\value{
parsed results from geocoder
}
\description{
Workhorse function for geocoding
}
