% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_cascade.R
\name{geo_cascade}
\alias{geo_cascade}
\title{Geocode a street address}
\usage{
geo_cascade(address, verbose = FALSE, lat = Lat, long = lng)
}
\arguments{
\item{address}{Single line address. Street must be included.}

\item{verbose}{logical. If TRUE outputs logs.}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}
}
\value{
Latitude and Longitude Coordinates in tibble format
}
\description{
First attempts to use the US Census Geocoder (geo_census) method
and then Nominatim/OSM (geo_osm) method as a backup
Returns latitude and longitude coordinates.
}
\examples{
geo_cascade("1600 Pennsylvania Ave Washington, DC")
geo_cascade("Paris, France")
}
