% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_census.R
\name{geo_census}
\alias{geo_census}
\title{Geocode street addresses}
\usage{
geo_census(address, verbose = FALSE, lat = lat, long = long,
  benchmark = 4,
  API_URL = "https://geocoding.geo.census.gov/geocoder/locations/onelineaddress?")
}
\arguments{
\item{address}{single line address. Street must be included.}

\item{verbose}{logical. If TRUE outputs logs.}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{benchmark}{parameter for the US Census Geocoder}

\item{API_URL}{URL of Census API}
}
\description{
Obtain latitude and longitude coordinates from an address using the US Census geocoder.
Only works for addresses within the US. Addresses must also be at the street
level (ie. 60 Main St. Pawnee, IN not Pawnee, IN).
}
