% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_methods.R
\name{geo_census}
\alias{geo_census}
\alias{geo_osm}
\alias{geo_geocodio}
\alias{geo_iq}
\alias{geo_google}
\alias{geo_opencage}
\alias{geo_mapbox}
\alias{geo_here}
\title{Convenience functions for calling the \code{\link{geo}} function
with a specified method}
\usage{
geo_census(...)

geo_osm(...)

geo_geocodio(...)

geo_iq(...)

geo_google(...)

geo_opencage(...)

geo_mapbox(...)

geo_here(...)
}
\arguments{
\item{...}{arguments to be passed to the \code{\link{geo}} function}
}
\description{
Convenience functions for calling the \code{\link{geo}} function
with a specified method
}
