% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_methods.R
\name{geo_census}
\alias{geo_census}
\alias{geo_osm}
\title{Conveniance functions for calling the \code{\link{geo}} function
with a specified method.}
\usage{
geo_census(...)

geo_osm(...)
}
\description{
Conveniance functions for calling the \code{\link{geo}} function
with a specified method.
}
