% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_census.R
\name{geo_census}
\alias{geo_census}
\title{Geocode a street address using the US Census Geocoder API}
\usage{
geo_census(address, verbose = FALSE, lat = latitude,
  long = longitude,
  API_URL = "https://geocoding.geo.census.gov/geocoder/locations/onelineaddress?")
}
\arguments{
\item{address}{Single line address. Street must be included.}

\item{verbose}{logical. If TRUE outputs logs.}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{API_URL}{URL of Census API}
}
\description{
Returns latitude and longitude coordinates. Only works for US addresses.
Address must be at the street level (ie. 60 Main St. Pawnee, IN not Pawnee, IN)
}
