% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_combine.R
\name{geo_combine}
\alias{geo_combine}
\title{Combine multiple geocoding queries}
\usage{
geo_combine(
  queries,
  global_params = list(),
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  ...
)
}
\arguments{
\item{queries}{list of lists parameter. Each list contains parameters for a query
(ie. \code{list(list(method = 'osm'), list(method = 'census'), ...)})}

\item{global_params}{list parameter. Contains arguments that should be used for all queries.
(ie. \code{list(full_results = TRUE, unique_only = TRUE)})}

\item{address}{single line address (ie. '1600 Pennsylvania Ave NW, Washington, DC').
Do not combine with the address component arguments below
(street, city, county, state, postalcode, country).}

\item{street}{street address (ie. '1600 Pennsylvania Ave NW')}

\item{city}{city (ie. 'Tokyo')}

\item{county}{county (ie. 'Jefferson')}

\item{state}{state (ie. 'Kentucky')}

\item{postalcode}{postalcode (zip code if in the United States)}

\item{country}{country (ie. 'Japan')}

\item{...}{arguments passed to the \link{geocode_combine} function}
}
\value{
tibble (dataframe)
}
\description{
Replaces method = 'cascade'
}
\examples{
\donttest{
geo_combine(address = c(""), list(list(method = 'census')))
}
}
\seealso{
\link{geocode_combine} \link{geo} \link{geocode}
}
