% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_osm.R
\name{geo_osm}
\alias{geo_osm}
\title{Geocode addresses}
\usage{
geo_osm(
  address,
  lat = lat,
  long = long,
  min_time = 1,
  verbose = FALSE,
  debug = FALSE,
  server = "http://nominatim.openstreetmap.org"
)
}
\arguments{
\item{address}{single line address}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{min_time}{Minimum time (in seconds) for the query to take. 
Defaults to 1 second per the Nominatim usage policy. 
Be cautious of setting this value lower as the usage limit may be reached.}

\item{verbose}{logical. If TRUE outputs logs.}

\item{debug}{logical. If TRUE outputs debugging logs}

\item{server}{URL to OSM Nominatim server}
}
\value{
latitude and longitude coordinates in tibble format
}
\description{
Obtains latitude and longitude coordinates from an address using
the Nominatim (OSM) geocoder service. Can be used with non-US
or non-street level addresses unlike the Census geocoder.
}
\details{
WARNING - This service has a usage limit and it will return
missing coordinates once the usage limit is reached. The min_time
argument defaults to 1 second to avoid this usage limit being reached.

Query adapted from tmaptools::geocode_OSM.
}
\examples{
\donttest{
geo_osm("1600 Pennsylvania Ave Washington, DC")
geo_osm('20 downing st london',debug=T)
geo_osm("Paris, France",verbose=TRUE)
}
}
\references{
tmaptools: \url{https://github.com/mtennekes/tmaptools}

Nominatim usage policy: \url{https://operations.osmfoundation.org/policies/nominatim}

Nominatim API referene: \url{https://nominatim.org/release-docs/develop/api/Search/}
}
