% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_osm.R
\name{geo_osm}
\alias{geo_osm}
\title{Geocode an address using the Nominatim (OSM) Geocoder API}
\usage{
geo_osm(address, verbose = FALSE, lat = latitude, long = longitude)
}
\arguments{
\item{address}{Single line address. Street must be included.}

\item{verbose}{logical. If TRUE outputs logs.}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}
}
\value{
Latitude and Longitude Coordinates in Tibble format
}
\description{
Returns latitude and longitude coordinates.
}
\examples{
geo_osm("1600 Pennsylvania Ave Washington, DC")
geo_osm("Paris, France")
}
