% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_osm.R
\name{geo_osm}
\alias{geo_osm}
\title{Geocode addresses}
\usage{
geo_osm(address, verbose = FALSE, lat = lat, long = long)
}
\arguments{
\item{address}{single line address}

\item{verbose}{logical. If TRUE outputs logs.}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}
}
\value{
latitude and longitude Coordinates in tibble format
}
\description{
Obtains latitude and longitude coordinates from an address using
the Nominatim (OSM) geocoder service. Can be used with non-US
or non-street level addresses unlike the Census geocoder. This
function calls the geocode_OSM function from the tmaptools package.
}
\details{
WARNING - This service has a usage limit and it will return
missing coordinates once the usage limit is reached.
}
\examples{
\dontrun{
geo_osm("1600 Pennsylvania Ave Washington, DC")
geo_osm("Paris, France")
}
}
