% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode addresses in a dataframe}
\usage{
geocode(
  .tbl,
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  lat = lat,
  long = long,
  return_addresses = FALSE,
  ...
)
}
\arguments{
\item{.tbl}{dataframe containing addresses

Names of columns in the dataframe:}

\item{address}{single line street address column name. Do not combine with 
address component arguments (street, city, county, state, postalcode, country)}

\item{street}{street address column name}

\item{city}{city column name}

\item{county}{county column name}

\item{state}{state column name}

\item{postalcode}{postalcode column name (zip code if in the United States)}

\item{country}{country column name}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{return_addresses}{if TRUE then addresses with standard names will be returned
This is defaulted to FALSE because the address fields are already in the input dataset}

\item{...}{arguments passed to the \code{\link{geo}} function}
}
\value{
input dataframe (.tbl) with geocoder results appended as columns
}
\description{
Takes a dataframe containing addresses as a input. It returns
the dataframe with latitude and longitude coordinate columns
using a user specified geocoder service. This function passes
all additional parameters (...) to the \code{\link{geo}} function
so you can refer to its documentation for more details on possible
arguments.
}
\details{
See example usage in \code{vignette("tidygeocoder")}
}
\examples{
\donttest{
library(dplyr)
sample_addresses[1:3,] \%>\% geocode(addr)

louisville[1:2, ] \%>\% geocode(street = street, city = city, state = state,
  postalcode = zip)

sample_addresses[8:9,] \%>\% geocode(addr, method='osm',lat='lattes', long='longos')

sample_addresses[1:3,] \%>\% geocode(addr, method='cascade', lat=latitude, long=longitude)
}
}
