% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode a street address}
\usage{
geocode(.tbl, address, method = "Census", ...)
}
\arguments{
\item{.tbl}{Dataframe}

\item{address}{Single line address. Street must be included.}

\item{method}{What geocoder api do you want to use? 'OSM' or 'Census'}

\item{...}{arguments supplied to the relevant geocoder function
(see \code{\link{geo_census}} (Census) and \code{\link{geo_osm}} (OSM))}
}
\value{
Latitude and Longitude Coordinates in tibble format
}
\description{
Returns latitude and longitude as two columns appended to
the inputted dataframe
}
\examples{
\dontrun{
sample_addresses \%>\% geocode(addr)
}

}
