% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode addresses in a dataframe and append latitude and longitude as columns}
\usage{
geocode(.tbl, address, method = "census", lat = lat, long = lng, ...)
}
\arguments{
\item{.tbl}{Dataframe}

\item{address}{Single line address. Street must be included.}

\item{method}{the geocoder function/service that you want to use
\itemize{
  \item "census": \code{\link{geo_census}}
  \item "osm": \code{\link{geo_osm}}
}}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{...}{arguments supplied to the relevant geocoder function}
}
\value{
input dataframe (.tbl) with latitude and longitude fields appended
}
\description{
Returns latitude and longitude as two columns appended to
the inputted dataframe
}
\examples{
\dontrun{
sample_addresses \%>\% geocode(addr)

sample_addresses \%>\% geocode(addr,method='osm',lat=latitude,long=longitude)
}
}
