% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode addresses in a dataframe}
\usage{
geocode(
  .tbl,
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  method = "census",
  lat = lat,
  long = long,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{.tbl}{dataframe

Names of columns in the dataframe:}

\item{address}{single line street address}

\item{street}{street address}

\item{city}{city}

\item{county}{county}

\item{state}{state}

\item{postalcode}{postalcode (zip code if in the United States)}

\item{country}{country}

\item{method}{the geocoder function you want to use
\itemize{
  \item "census": only for US street level addresses
  \item "osm": worldwide coverage but has a usage limit
}}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{verbose}{toggle verbose output}

\item{...}{arguments passed to the \code{\link{geo}} function}
}
\value{
input dataframe (.tbl) with latitude and longitude fields appended
}
\description{
Takes a dataframe containing addresses as a input. Returns
the dataframe with latitude and longitude coordinate columns
using a user specified geocoder function.
}
\details{
See example usage in \code{vignette("tidygeocoder")}
}
\examples{
\donttest{
library(dplyr)
sample_addresses \%>\% geocode(addr)

sample_addresses \%>\% geocode(addr,method='cascade',lat=latitude,long=longitude)
}
}
