% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Tidyverse-style interface for geocoding addresses}
\usage{
geocode(.tbl, address, method = "census", lat = latitude,
  long = longitude, ...)
}
\arguments{
\item{.tbl}{Dataframe}

\item{address}{Single line address. Street must be included.}

\item{method}{the geocoder function/service that you want to use
\itemize{
  \item "census": \code{\link{geo_census}}
  \item "osm": \code{\link{geo_osm}}
  \item "cascade": \code{\link{geo_cascade}} (first tries to use census then tries osm)
}}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{...}{arguments supplied to the relevant geocoder function}
}
\value{
input dataframe (.tbl) with latitude and longitude fields appended
}
\description{
Returns latitude and longitude as two columns appended to
the inputted dataframe.
See example usage in vignette("geocode")
}
\examples{
\dontrun{
sample_addresses \%>\% geocode(addr)

sample_addresses \%>\% geocode(addr,method='cascade',lat=latitude,long=longitude)
}
}
