% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode street addresses in a tibble}
\usage{
geocode(.tbl, address, method = "census", lat = lat, long = long,
  ...)
}
\arguments{
\item{.tbl}{Dataframe}

\item{address}{Single line address.
\itemize{
  \item "census": \code{\link{geo_census}} - can only handle US street level addresses
  \item "osm": \code{\link{geo_osm}} - more versatile than Census but has a usage limit
  \item "cascade": \code{\link{geo_cascade}} - first tries to use census then tries osm
}}

\item{lat}{name of latitude field}

\item{long}{name of longitude field}

\item{...}{arguments supplied to the relevant geocoder function}
}
\value{
input dataframe (.tbl) with latitude and longitude fields appended
}
\description{
Returns latitude and longitude as two columns appended to
the inputted dataframe.
See example usage in vignette("geocode")
}
\examples{
\dontrun{
sample_addresses \%>\% geocode(addr)

sample_addresses \%>\% geocode(addr,method='cascade',lat=latitude,long=longitude)
}
}
