% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode addresses in a dataframe}
\usage{
geocode(
  .tbl,
  address = NULL,
  street = NULL,
  city = NULL,
  county = NULL,
  state = NULL,
  postalcode = NULL,
  country = NULL,
  lat = lat,
  long = long,
  limit = 1,
  return_addresses = FALSE,
  unique_only = FALSE,
  ...
)
}
\arguments{
\item{.tbl}{dataframe containing addresses}

\item{address}{single line street address column name. Do not combine with
address component arguments (street, city, county, state, postalcode, country)}

\item{street}{street address column name}

\item{city}{city column name}

\item{county}{county column name}

\item{state}{state column name}

\item{postalcode}{postalcode column name (zip code if in the United States)}

\item{country}{country column name}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. lat or 'lat').}

\item{long}{longitude column name. Can be quoted or unquoted (ie. long or 'long').}

\item{limit}{maximum number of results to return per address. For many geocoder services
the maximum value for the limit parameter is 100. Pass \code{limit = NULL} to use
the default \code{limit} value of the selected geocoder service.
For batch geocoding, limit must be set to 1 (default) if \code{return_addresses = TRUE}
To use \code{limit > 1} or \code{limit = NULL} either \code{return_addresses} or \code{unique_only} must be set to TRUE.
Refer to \link{api_parameter_reference} for more details.}

\item{return_addresses}{return input addresses with results if TRUE. Note that
most services return the input addresses with \code{full_results = TRUE} and setting
return_addresses to FALSE does not prevent this.}

\item{unique_only}{only return results for unique inputs if TRUE}

\item{...}{arguments passed to the \link{geo} function}
}
\value{
tibble (dataframe)
}
\description{
Takes a dataframe containing addresses as an input and returns
the results from a specified geocoder service in a dataframe format using the
\link{geo} function. See example usage in \code{vignette("tidygeocoder")}.

This function passes all additional parameters (\code{...}) to the
\code{\link{geo}} function, so you can refer to its documentation for more details
on possible arguments.

Note that the arguments used for specifying address columns (address,
street, city, county, state, postalcode, and country) accept either
quoted or unquoted column names (ie. "address_col" and address_col are
both acceptable).
}
\examples{
\donttest{
library(dplyr)
sample_addresses[1:2,] \%>\% geocode(addr)

louisville[1:2,] \%>\% geocode(street = street, city = city, state = state,
  postalcode = zip)

sample_addresses[8:9,] \%>\% geocode(addr, method = 'osm',
  lat = 'lattes', long = 'longos')

sample_addresses[4:5,] \%>\% geocode(addr, method = 'cascade',
  lat = latitude, long = longitude)
}
}
\seealso{
\code{\link{geo}}
}
