% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_combine.R
\name{geocode_combine}
\alias{geocode_combine}
\title{Combine multiple geocoding queries}
\usage{
geocode_combine(
  .tbl,
  queries,
  global_params = list(),
  query_names = NULL,
  return_list = FALSE,
  cascade = TRUE,
  lat = lat,
  long = long
)
}
\arguments{
\item{.tbl}{dataframe containing addresses}

\item{queries}{list of lists parameter. Each list contains parameters for a query
(ie. \code{list(list(method = 'osm'), list(method = 'census'), ...)})}

\item{global_params}{list parameter. Contains parameters to be used for all queries.
(ie. \code{list(full_results = TRUE, unique_only = TRUE)})}

\item{query_names}{an optional vector of names to use for labeling the results of each query.
These names are either placed in a \code{query} column if \code{cascade = TRUE}.
If null then the \code{method} parameter values will be used as names.}

\item{return_list}{if TRUE then results from each service will be returned as separate
items in a named list. If FALSE (default) then all results will be combined into a
single dataframe.}

\item{cascade}{if TRUE (default) then only addresses that are not found by a geocoding
service will be attempted by subsequent queries.
If FALSE then all queries will attempt to geocode all addresses.}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. lat or 'lat').}

\item{long}{longitude column name. Can be quoted or unquoted (ie. long or 'long').}
}
\value{
tibble (dataframe)
}
\description{
Executes multiple geocoding queries on a dataframe input and combines
the results. To use a character vector input instead, see the \link{geo_combine} function.
}
\examples{
\donttest{
geocode_combine(sample_addresses,
 queries = list(list(method = 'census'), list(method = 'osm')),
 global_params = list(address = 'addr'), cascade = TRUE)

 geocode_combine( 
     tibble::tribble(
     ~res_street_address, ~res_city_desc, ~state_cd, ~zip_code,
     "624 W DAVIS ST #1D",   "BURLINGTON", "NC", 27215,
     "201 E CENTER ST #268", "MEBANE",     "NC", 27302,
     "100 Wall Street",      "New York",   "NY", 10005
     ),
     queries = list(
         list(method = 'census', mode = 'batch'),
         list(method = 'census', mode = 'single')
      ),
     global_params = list(street = 'res_street_address', 
       city = 'res_city_desc', state = 'state_cd', postalcode = 'zip_code'),
     query_names = c('census batch', 'census single')
 )
}
}
\seealso{
\link{geo_combine} \link{geo} \link{geocode}
}
