% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_combine.R
\name{geocode_combine}
\alias{geocode_combine}
\title{Combine multiple geocoding queries}
\usage{
geocode_combine(
  .tbl,
  queries,
  global_params = list(),
  query_names = NULL,
  return_list = FALSE,
  cascade = TRUE,
  lat = lat,
  long = long
)
}
\arguments{
\item{.tbl}{dataframe containing addresses}

\item{queries}{a list of queries, each provided as a list of parameters. The queries are
executed in the order provided.
(ex. \code{list(list(method = 'osm'), list(method = 'census'), ...)})}

\item{global_params}{a list of parameters to be used for all queries
(ex. \code{list(address = 'address', full_results = TRUE)})}

\item{query_names}{an optional vector of names to use for labeling the results of each query.
These names are either placed in a \code{query} column if \code{cascade = TRUE}.
If not specified (NULL) then the \code{method} parameter values will be used to label the query results.
If the same \code{method} is used in multiple queries, then a number will be added to the method name
to reflect the order of the queries.}

\item{return_list}{if TRUE then results from each service will be returned as separate
items in a named list. If FALSE (default) then all results will be combined into a
single dataframe.}

\item{cascade}{if TRUE (default) then only addresses that are not found by a geocoding
service will be attempted by subsequent queries.
If FALSE then all queries will attempt to geocode all addresses.}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. lat or 'lat').}

\item{long}{longitude column name. Can be quoted or unquoted (ie. long or 'long').}
}
\value{
tibble (dataframe)
}
\description{
Executes multiple geocoding queries on a dataframe input and combines
the results. To use a character vector input instead, see the \link{geo_combine} function.
}
\examples{
\donttest{

library(dplyr, warn.conflicts = FALSE)

sample_addresses \%>\%
  geocode_combine(
    queries = list(list(method = 'census'), list(method = 'osm')),
    global_params = list(address = 'addr'), cascade = TRUE)

more_addresses <- tibble::tribble(
     ~street_address, ~city, ~state,        ~zip_cd,
     "624 W DAVIS ST #1D",   "BURLINGTON", "NC", 27215,
     "201 E CENTER ST #268", "MEBANE",     "NC", 27302,
     "100 Wall Street",      "New York",   "NY", 10005,
     "Bucharest",            NA,           NA,   NA
     )
 
 more_addresses \%>\%        
   geocode_combine( 
     queries = list(
         list(method = 'census', mode = 'batch'),
         list(method = 'census', mode = 'single'),
         list(method = 'osm')
      ),
     global_params = list(street = 'street_address', 
       city = 'city', state = 'state', postalcode = 'zip_cd'),
     query_names = c('census batch', 'census single', 'osm')
   )
   
 more_addresses \%>\%
   geocode_combine( 
     queries = list(
         list(method = 'census', mode = 'batch', street = 'street_address', 
       city = 'city', state = 'state', postalcode = 'zip_cd'),
         list(method = 'arcgis', address = 'street_address')
      ),
     cascade = FALSE,
     return_list = TRUE
   )
}
}
\seealso{
\link{geo_combine} \link{geo} \link{geocode}
}
