% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_loop.R
\name{geocode_loop}
\alias{geocode_loop}
\title{New function to replace method = 'cascade'}
\usage{
geocode_loop(
  .tbl,
  queries,
  common_params = list(),
  stack = TRUE,
  cascade = TRUE,
  lat = lat,
  long = long
)
}
\arguments{
\item{.tbl}{dataframe containing addresses}

\item{queries}{list of lists parameter. Each list contains parameters for a query
(ie. \code{list(list(method = 'osm'), list(method = 'census'), ...)})}

\item{common_params}{list parameter. Contains arguments that should be used for all queries.
(ie. \code{list(full_results = TRUE, unique_only = TRUE)})}

\item{stack}{if TRUE then a single dataframe will be returned. When stack = FALSE
then a list of dataframes, one per each query, will be returned.}

\item{cascade}{if TRUE then only addresses that are not found will be attempted by
the following query. If FALSE then all queries will attempt to geocode all addresses.}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. lat or 'lat').}

\item{long}{longitude column name. Can be quoted or unquoted (ie. long or 'long').}
}
\value{
tibble (dataframe)
}
\description{

}
