% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_factory.R
\name{get_api_query}
\alias{get_api_query}
\title{Function for creating api queries}
\usage{
get_api_query(
  method_name,
  generic_parameters = list(),
  custom_api_parameters = list()
)
}
\description{
Function for creating api queries
}
