% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_factory.R
\name{get_api_query}
\alias{get_api_query}
\title{Construct a Geocoder API Query}
\usage{
get_api_query(method, generic_parameters = list(), custom_parameters = list())
}
\arguments{
\item{method}{method name (ie. 'census')}

\item{generic_parameters}{universal 'generic' parameters}

\item{custom_parameters}{custom api-specific parameters}
}
\value{
named list of parameters
}
\description{
The geocoder API query is created using universal "generic" parameters
and optional api-specific "custom" parameters. Generic parameters
are converted into api parameters using the  \code{\link{api_parameter_reference}} 
dataset. 

The \code{\link{query_api}} function executes the queries created 
by this function.
}
\seealso{
\code{\link{query_api}} \code{\link{geo}}
}
