% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{louisville}
\alias{louisville}
\title{Louisville, Kentucky street addresses}
\format{
A tibble dataframe with component street addresses
\describe{
 \item{street}{Description of the address}
 \item{city}{Single line address}
 \item{state}{state}
 \item{zip}{zip code}
}
}
\source{
Downloaded from \href{http://results.openaddresses.io/sources/us/ky/jefferson}{OpenAddresses.io}
 on June 1st 2020
}
\usage{
louisville
}
\description{
Louisville, Kentucky street addresses
}
\keyword{datasets}
