% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_time_reference}
\alias{min_time_reference}
\title{The minimum number of seconds required per query to comply with usage restrictions}
\format{
A tibble dataframe
\describe{
 \item{method}{Geocoder service name}
 \item{min_time}{The minimum number of seconds required per query to comply with usage restrictions}
 \item{description}{A description of the usage rate restriction}
}
}
\usage{
min_time_reference
}
\description{
The minimum number of seconds required per query to comply with usage restrictions
}
\keyword{datasets}
