% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_time_reference}
\alias{min_time_reference}
\title{Minimum time required per query}
\format{
A tibble dataframe
\describe{
\item{method}{Geocoding service name}
\item{min_time}{The minimum number of seconds required per query to comply with usage restrictions}
\item{description}{A description of the usage rate restriction}
}
}
\usage{
min_time_reference
}
\description{
The \link{geo} and \link{reverse_geo} functions use this dataset
to set the maximum query rate for each geocoding service.
This rate is based on the usage restriction policies for
each geocoding service.
}
\details{
Links to the usage policies of each geocoding service are below:
\itemize{
\item \href{https://operations.osmfoundation.org/policies/nominatim/}{Nominatim}, - \href{https://www.census.gov/programs-surveys/geography/technical-documentation/complete-technical-documentation/census-geocoder.html}{US Census}, - \href{https://developers.arcgis.com/pricing/}{ArcGIS}, - \href{https://www.geocod.io/pricing/}{Geocodio}, - \href{https://locationiq.com/pricing}{Location IQ}, - \href{https://developers.google.com/maps/documentation/geocoding/usage-and-billing}{Google}, - \href{https://opencagedata.com/pricing}{OpenCage}, - \href{https://www.mapbox.com/pricing/}{Mapbox}, - \href{https://developer.here.com/pricing}{HERE}, - \href{https://developer.tomtom.com/store/maps-api}{TomTom}, - \href{https://developer.mapquest.com/plans}{MapQuest}, - \href{https://docs.microsoft.com/en-us/bingmaps/spatial-data-services/geocode-and-data-source-limits}{Bing}, - \href{https://www.geoapify.com/term-and-conditions}{Geoapify}
}
}
\seealso{
\link{geo} \link{reverse_geo}
}
\keyword{datasets}
