% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_factory.R
\name{query_api}
\alias{query_api}
\title{Get raw results back from an API}
\usage{
query_api(api_url, query_parameters, content_encoding = "UTF-8")
}
\arguments{
\item{api_url}{Base URL of the API. query parameters are appended to this}

\item{query_parameters}{api query parameters in the form of a named list}

\item{content_encoding}{Encoding to be used for parsing content.}
}
\value{
raw results from the query
Census uses "ISO-8859-1", all other services use "UTF-8"
}
\description{
Get raw results back from an API
}
