% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geo.R
\name{reverse_geo}
\alias{reverse_geo}
\title{lat, long = inputs
address = name of address column}
\usage{
reverse_geo(
  lat,
  long,
  address = "address",
  method = "osm",
  limit = 1,
  api_url = NULL,
  full_results = FALSE,
  unique_only = FALSE,
  flatten = TRUE,
  verbose = FALSE,
  no_query = FALSE,
  custom_query = list(),
  geocodio_v = 1.6,
  iq_region = "us",
  param_error = TRUE
)
}
\description{
lat, long = inputs
address = name of address column
}
