% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse geocode coordinates in a dataframe}
\usage{
reverse_geocode(
  .tbl,
  lat,
  long,
  address = address,
  limit = 1,
  return_coords = FALSE,
  unique_only = FALSE,
  ...
)
}
\arguments{
\item{.tbl}{dataframe containing coordinates}

\item{lat}{latitude column name (input data). Can be quoted or unquoted (ie. lat or 'lat').}

\item{long}{longitude column name (input data). Can be quoted or unquoted (ie. long or 'long').}

\item{address}{address column name (output data). Can be quoted or unquoted (ie. addr or 'addr').}

\item{limit}{maximum number of results to return per coordinate. For many geocoder services
the maximum value for the limit parameter is 100. Pass \code{limit = NULL} to use
the default \code{limit} value of the selected geocoder service.
For batch geocoding, limit must be set to 1 (default) if \code{return_coords = TRUE}
To use \code{limit > 1} or \code{limit = NULL} either \code{return_coords} or \code{unique_only} must be set to TRUE.
Refer to \link{api_parameter_reference} for more details.}

\item{return_coords}{if TRUE then only the geocoder results and input coordinate data will be returned.
if FALSE then the input dataset's columns will also be included.}

\item{unique_only}{if TRUE then only unique coordinates and results will be returned.
The input dataframe's format is not preserved. Coordinates will also be returned if
TRUE (overrides return_coords argument).}

\item{...}{arguments passed to the \link{reverse_geo} function}
}
\value{
tibble (dataframe)
}
\description{
Takes a dataframe containing coordinates (latitude and longitude) and returns
the reverse geocoding query results from a specified service by using the
\code{\link{reverse_geo}} function. See example usage in \code{vignette("tidygeocoder")}.

This function passes all additional parameters (\code{...}) to the
\code{\link{reverse_geo}} function, so you can refer to its documentation for more details
on possible arguments.
}
\examples{
\donttest{
library(tibble)
reverse_geocode(
  tibble(
    latitude = c(38.895865, 43.6534817),
    longitude = c(-77.0307713,-79.3839347)
  ),
  lat = latitude,
  long = longitude,
  method = 'osm',
  full_results = TRUE,
  verbose = TRUE
)
}
}
\seealso{
\link{reverse_geo}
}
