% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Reverse geocode coordinates in a dataframe}
\usage{
reverse_geocode(
  .tbl,
  lat,
  long,
  address = address,
  return_coords = FALSE,
  unique_only = FALSE,
  ...
)
}
\arguments{
\item{.tbl}{dataframe containing coordinates}

\item{lat}{latitude column name. Can be quoted or unquoted (ie. lat or 'lat').}

\item{long}{longitude column name. Can be quoted or unquoted (ie. long or 'long').}

\item{address}{address column name.}

\item{return_coords}{if TRUE then coordinates with standard names will be returned
This is defaulted to FALSE because the coordinates are already in the input dataset}

\item{unique_only}{if TRUE then only unique coordinates and results will be returned. 
The input dataframe's format is not preserved. Coordinates will also be returned if 
TRUE (overrides return_coords argument).}

\item{...}{arguments passed to the \code{\link{reverse_geo}} function}
}
\value{
input dataframe (.tbl) with geocoder results appended as columns
}
\description{
Takes a dataframe containing latitude, longitude coordinates as inputs and returns 
the dataframe results from a specified geocoder service by using the
\code{\link{reverse_geo}} function. See example usage in \code{vignette("tidygeocoder")}.

This function passes all additional parameters (\code{...}) to the 
\code{\link{reverse_geo}} function, so you can refer to its documentation for more details
on possible arguments.
}
\examples{
\donttest{
library(tibble)
reverse_geocode(
  tibble(
    latitude = c(38.895865, 43.6534817),
    longitude = c(-77.0307713,-79.3839347)
  ),
  lat = latitude,
  long = longitude,
  method = 'osm',
  full_results = TRUE,
  verbose = TRUE
)
}
}
\seealso{
\code{\link{reverse_geo}} \code{\link{api_parameter_reference}}
}
