% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_addresses}
\alias{sample_addresses}
\title{Some sample addresses for testing}
\format{
A tibble dataframe with single line addresses
\describe{
 \item{name}{Description of the address}
 \item{addr}{Single line address}
}
}
\usage{
sample_addresses
}
\description{
Some sample addresses for testing
}
\keyword{datasets}
