% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{tidygeocoder: Tidyverse-style interface for geocoding}
\description{
This package provides an intuitive tidyverse-style interface for geocoding services. The currently supported services are the US Census geocoder and Nominatim (OSM).
}
\details{
A tidyverse-style geocoder interface for R. Utilizes US Census and Nominatim (OSM) geocoder services.

The primary function to use is \code{\link{geocode}} which can call
\code{\link{geo_census}}, \code{\link{geo_osm}}, or \code{\link{geo_cascade}} depending on the method argument.
See vignette("geocode") for more details
}
\references{
\itemize{
  \item Tidygeocoder package: \url{https://github.com/jessecambon/tidygeocoder}
  \item US Census Geocoder: \url{https://geocoding.geo.census.gov/}
  \item Nominatim (OSM): \url{https://nominatim.openstreetmap.org/} and \url{https://nominatim.org/}
  \item Tmaptools package (used for OSM geocoding): \url{https://cran.r-project.org/web/packages/tmaptools/index.html}
  \item Dplyr (Tidyverse): \url{https://dplyr.tidyverse.org/}
}
}
\keyword{internal}
