% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{tidygeocoder: Tidyverse-style interface for geocoding.}
\description{
Functions for obtaining latitude and longitude from street addresses
}
\details{
Functions for geocoding (obtaining latitudes and longitude
coordinates from addreses)
}
\keyword{internal}
