% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygeocoder.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{\pkg{tidygeocoder} makes getting data from geocoder services easy.}
\description{
The primary functions to use are \code{\link{geocode}} and \code{\link{geo}}.
Refer to the documentation on the method argument of the \code{\link{geo}} function
for more details on the available geocoder services. Also see \code{vignette("tidygeocoder")} 
for example usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jessecambon.github.io/tidygeocoder/}
  \item \url{https://github.com/jessecambon/tidygeocoder}
  \item Report bugs at \url{https://github.com/jessecambon/tidygeocoder/issues}
}

}
\author{
\strong{Maintainer}: Jesse Cambon \email{jesse.cambon@gmail.com} (\href{https://orcid.org/0000-0001-6854-1514}{ORCID})

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})
  \item Christopher Belanger \email{christopher.a.belanger@gmail.com}
  \item Daniel Possenriede \email{possenriede+r@gmail.com} (\href{https://orcid.org/0000-0002-6738-9845}{ORCID})
}

}
\keyword{internal}
