% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{tidygeocoder: Tidyverse-style interface for geocoding}
\description{
This package provides an intuitive interface for using geocoding services. The currently supported services are the US Census geocoder and Nomanatim (OSM).
}
\details{
Functions for geocoding (obtaining latitudes and longitude
coordinates from addreses). US Census Geocoder and Nominatim services are supported.

The primary function to use is \code{\link{geocode}} which can call either
\code{\link{geo_census}} or \code{\link{geo_osm}} depending on the method argument.
See vignette("geocode") for more details
}
\references{
\itemize{
  \item Tidygeocoder Github: \url{https://github.com/jessecambon/tidygeocoder}
  \item Dplyr (Tidyverse): \url{https://dplyr.tidyverse.org/}
  \item US Census Geocoder: \url{https://geocoding.geo.census.gov/}
  \item Nominatim (OSM): \url{https://nominatim.openstreetmap.org/}
}
}
\keyword{internal}
