% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygeocoder.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{\pkg{tidygeocoder} makes getting data from geocoder services easy.}
\description{
The primary functions to use are \code{\link{geocode}} and \code{\link{geo}}.
Refer to the documentation on the method argument of the \code{\link{geo}} function
for more details on the available geocoder services. Also see \code{vignette("tidygeocoder")} 
for example usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jessecambon.github.io/tidygeocoder/}
  \item \url{https://github.com/jessecambon/tidygeocoder}
  \item Report bugs at \url{https://github.com/jessecambon/tidygeocoder/issues}
}

}
\author{
\strong{Maintainer}: Jesse Cambon \email{jesse.cambon@gmail.com}

Authors:
\itemize{
  \item Christopher Belanger \email{christopher.a.belanger@gmail.com}
}

}
\keyword{internal}
