% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygeocoder.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{The tidygeocoder package makes getting data from geocoder services easy.}
\description{
The \link{geocode} and \link{geo} functions are for forward geocoding while
the \link{reverse_geocode} and \link{reverse_geo} functions are for reverse geocoding.
Refer to the documentation on the \code{method} argument in the \link{geo} and \link{reverse_geo} functions
for more details on the available geocoding services. Also see \code{vignette("tidygeocoder")}
for example usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jessecambon.github.io/tidygeocoder/}
  \item \url{https://github.com/jessecambon/tidygeocoder}
  \item Report bugs at \url{https://github.com/jessecambon/tidygeocoder/issues}
}

}
\author{
\strong{Maintainer}: Jesse Cambon \email{jesse.cambon@gmail.com} (\href{https://orcid.org/0000-0001-6854-1514}{ORCID})

Authors:
\itemize{
  \item Diego Hernangómez \email{diego.hernangomezherrero@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})
  \item Christopher Belanger \email{christopher.a.belanger@gmail.com} (\href{https://orcid.org/0000-0003-2070-5721}{ORCID})
  \item Daniel Possenriede \email{possenriede+r@gmail.com} (\href{https://orcid.org/0000-0002-6738-9845}{ORCID})
}

Other contributors:
\itemize{
  \item Otto Hansen \email{otto@uchicago.edu} (\href{https://orcid.org/0000-0002-4618-5667}{ORCID}) [contributor]
}

}
\keyword{internal}
