% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygeocoder.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{\pkg{tidygeocoder} makes getting data from geocoder services easy.}
\description{
The primary functions to use are \code{\link{geocode}} and \code{\link{geo}}.
The \code{\link{geocode}} function calls \code{\link{geo}}.
See \code{vignette("tidygeocoder")} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jessecambon.github.io/tidygeocoder}
  \item \url{https://github.com/jessecambon/tidygeocoder}
  \item Report bugs at \url{https://github.com/jessecambon/tidygeocoder/issues}
}

}
\author{
\strong{Maintainer}: Jesse Cambon \email{jesse.cambon@gmail.com}

}
\keyword{internal}
