% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tidygeocoder-package}
\alias{tidygeocoder}
\alias{tidygeocoder-package}
\title{tidygeocoder: Tidyverse-Style Interface for Geocoding}
\description{
An intuitive tidyverse-style interface for geocoding. Obtains latitude and longitude coordinates in tibble format from addresses. The currently supported services are the US Census geocoder and Nominatim (OSM).
}
\details{
The primary function to use is \code{\link{geocode}} which can call
\code{\link{geo_census}}, \code{\link{geo_osm}}, or \code{\link{geo_cascade}} depending on the method argument.
See \code{vignette("tidygeocoder")} for details.
}
\references{
\itemize{
  \item US Census Geocoder: \url{https://geocoding.geo.census.gov/}
  \item Nominatim (OSM): \url{https://nominatim.openstreetmap.org/} and \url{https://nominatim.org/}
  \item tmaptools package (used for OSM geocoding): \url{https://cran.r-project.org/package=tmaptools}
  \item dplyr: \url{https://dplyr.tidyverse.org/}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jessecambon/tidygeocoder}
  \item Report bugs at \url{https://github.com/jessecambon/tidygeocoder/issues}
}

}
\keyword{internal}
