% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_handling.R
\name{unpackage_addresses}
\alias{unpackage_addresses}
\title{Function for unpackaging and RE-deduping addresses
so that we can return them in the same order that they were passed}
\usage{
unpackage_addresses(package, results)
}
\description{
Function for unpackaging and RE-deduping addresses
so that we can return them in the same order that they were passed
}
